requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-ollama
chromadb
python-dotenv
